+++ Model/ResourceModel/PreviewTemplate/Collection.php
@@ -3,8 +3,6 @@
 namespace Overdose\PreviewEmail\Model\ResourceModel\PreviewTemplate;
 
 use Magento\Framework\Model\ResourceModel\Db\Collection\AbstractCollection;
-use Overdose\PreviewEmail\Model\PreviewTemplate as Model;
-use Overdose\PreviewEmail\Model\ResourceModel\PreviewTemplate as Resource;
 
 /**
  * Class Collection
@@ -14,6 +12,9 @@ class Collection extends AbstractCollection
 {
     protected function _construct()
     {
-        $this->_init(Model::class, Resource::class);
+        $this->_init(
+            \Overdose\PreviewEmail\Model\PreviewTemplate::class,
+            \Overdose\PreviewEmail\Model\ResourceModel\PreviewTemplate::class
+        );
     }
 }